basic version done!

+++ app/Http/Controllers/Admin/PostController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\Admin;
 
+use App\Category;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
@@ -16,7 +17,7 @@ class PostController extends Controller
      */
     public function index()
     {
-        $posts = Post::all();
+        $posts = Post::with('category')->get();
         return view('admin.posts.index', compact('posts'));
     }
 
@@ -27,7 +28,8 @@ class PostController extends Controller
      */
     public function create()
     {
-        return view('admin.posts.create');
+        $categories = Category::all();
+        return view('admin.posts.create', compact('categories'));
     }
 
     /**
@@ -68,7 +70,11 @@ class PostController extends Controller
     public function show($id)
     {
         $post = Post::find($id);
-        return view('admin.posts.show', ['post' => $post]);
+        if ($post) {
+            return view('admin.posts.show', ['post' => $post]);
+        } else {
+            return abort('404');
+        }
     }
 
     /**
@@ -80,7 +86,16 @@ class PostController extends Controller
     public function edit($id)
     {
         $post= Post::find($id);
-        return view('admin.posts.edit', compact('post'));
+        $categories = Category::all();
+        $data = [
+            'post' => $post,
+            'categories' => $categories
+        ];
+        if ($post) {
+            return view('admin.posts.edit', $data);
+        } else {
+            return abort('404');
+        }
     }
 
     /**
@@ -124,7 +139,11 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
-        $post->delete();
-        return redirect()->route('admin.posts.index');
+        if ($post) {
+            $post->delete();
+            return redirect()->route('admin.posts.index');
+        } else {
+            return abort('404');
+        }
     }
 }
